% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example3Dtrees-data.R
\docType{data}
\name{example3Dtrees}
\alias{example3Dtrees}
\title{Examples of rooted 3D trees}
\format{
list
}
\usage{
data(example3Dtrees)
}
\description{
Contains the 3D models of 63 beans in phylo3D format. The single models
can be accessed with their IDs, e.g. "bean09" for the bean with ID 9. 
Note that all IDs from 01 to 66 are available except IDs 8, 29, and 60.
}
\examples{
data(example3Dtrees)
treeDbalance::example3Dtrees$bean09
}
\keyword{datasets}
